requirements.txt
python==3.8.6
django==3.2.2
Pillow==8.2.0
django-ckeditor==6.0.0